Redesign API for more use cases
This eliminates the dichotomy between synchronous and asynchronous
clients; clients are always synchronous. The presence of virtual
threads means that it is now trivial to cheaply treat anything
synchronous as asynchronous without having to provide an API for it.

The test suite now contains implementation examples for TCP, UDP,
and HTTP clients.

Fix: https://github.com/io7m/hibiscus/issues/2